requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.1
kaggle==1.6.11
psycopg2-binary==2.9.9 